removed abiguity of loads function
because you don't just import a function with such an unspecfic name like "loads". it has to be obvious what is being loaded by what or else the code becomes unreadable. If you want to midigate the (minimal) time spent by calling the function through module instead of directly, at least call it json_loads or loads_json or something.